Refactor navigation buttons and update styles
The code changes focused on refactoring the navigation buttons in the HTML page, introducing changes to simplify the button classes. Two new chevron SVG files have been added to represent navigation arrows. In the same line, The CSS for these navigation buttons was updated to reflect this change. Furthermore, the position of the 'artist-name-banner' div was adjusted.