Simplify song list structure and update css properties
This commit reduces unnecessary HTML divisions in the popular song list making the structure simpler and easier to maintain. At the same time, display properties have been updated in CSS, improving the visual flow and harmony of elements on the page. Adjustment in justify-content and align-items properties helps achieve a cleaner and more uniform interface.